+++ internal/cors/cors_test.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/gorilla/mux"
 	"github.com/stretchr/testify/require"
-	"github.com/xenitab/azad-kube-proxy/pkg/config"
+	"github.com/xenitab/azad-kube-proxy/internal/config"
 )
 
 func TestMiddleware(t *testing.T) {
