+++ internal/handlers/azure_test.go
