+++ internal/handlers/handlers.go
@@ -9,12 +9,12 @@ import (
 	"strings"
 
 	"github.com/go-logr/logr"
-	"github.com/xenitab/azad-kube-proxy/pkg/cache"
-	"github.com/xenitab/azad-kube-proxy/pkg/config"
-	"github.com/xenitab/azad-kube-proxy/pkg/health"
-	"github.com/xenitab/azad-kube-proxy/pkg/models"
-	"github.com/xenitab/azad-kube-proxy/pkg/user"
-	"github.com/xenitab/azad-kube-proxy/pkg/util"
+	"github.com/xenitab/azad-kube-proxy/internal/cache"
+	"github.com/xenitab/azad-kube-proxy/internal/config"
+	"github.com/xenitab/azad-kube-proxy/internal/health"
+	"github.com/xenitab/azad-kube-proxy/internal/models"
+	"github.com/xenitab/azad-kube-proxy/internal/user"
+	"github.com/xenitab/azad-kube-proxy/internal/util"
 	"github.com/xenitab/go-oidc-middleware/options"
 )
