+++ internal/proxy/proxy.go
@@ -15,15 +15,15 @@ import (
 
 	"github.com/go-logr/logr"
 	"github.com/gorilla/mux"
-	"github.com/xenitab/azad-kube-proxy/pkg/azure"
-	"github.com/xenitab/azad-kube-proxy/pkg/cache"
-	"github.com/xenitab/azad-kube-proxy/pkg/config"
-	"github.com/xenitab/azad-kube-proxy/pkg/cors"
-	"github.com/xenitab/azad-kube-proxy/pkg/handlers"
-	"github.com/xenitab/azad-kube-proxy/pkg/health"
-	"github.com/xenitab/azad-kube-proxy/pkg/metrics"
-	"github.com/xenitab/azad-kube-proxy/pkg/user"
-	"github.com/xenitab/azad-kube-proxy/pkg/util"
+	"github.com/xenitab/azad-kube-proxy/internal/azure"
+	"github.com/xenitab/azad-kube-proxy/internal/cache"
+	"github.com/xenitab/azad-kube-proxy/internal/config"
+	"github.com/xenitab/azad-kube-proxy/internal/cors"
+	"github.com/xenitab/azad-kube-proxy/internal/handlers"
+	"github.com/xenitab/azad-kube-proxy/internal/health"
+	"github.com/xenitab/azad-kube-proxy/internal/metrics"
+	"github.com/xenitab/azad-kube-proxy/internal/user"
+	"github.com/xenitab/azad-kube-proxy/internal/util"
 	"golang.org/x/sync/errgroup"
 )
