+++ internal/user/user.go
@@ -3,9 +3,9 @@ package user
 import (
 	"context"
 
-	"github.com/xenitab/azad-kube-proxy/pkg/azure"
-	"github.com/xenitab/azad-kube-proxy/pkg/config"
-	"github.com/xenitab/azad-kube-proxy/pkg/models"
+	"github.com/xenitab/azad-kube-proxy/internal/azure"
+	"github.com/xenitab/azad-kube-proxy/internal/config"
+	"github.com/xenitab/azad-kube-proxy/internal/models"
 )
 
 // ClientInterface ...
